+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # RELEASE NOTES
 ## 0.5.1 21-Dec-2020
 - Incorporated Inventory+ into Spell Better (converted to apply to spells)
+- Create new spell in special category (still shows in standard category however)
 ## 0.5.0 
 - Forked from ogl-5e-character-sheet???